+++ src/python/zensols/clojure.py
@@ -111,7 +111,7 @@ usage:
     @classmethod
     def eval_default(cls, code, context=None):
         """One shot convenience method to evaulate in the default namespace."""
-        inst = Clojure(namespace)
+        inst = Clojure()
         try:
             return inst.eval(code, context)
         finally:
@@ -120,4 +120,4 @@ usage:
     @classmethod
     def kill_server(cls):
         """Stop the Clojure p4j gateway server (careful!)."""
-        Clojure.eval_default('(zensols.py4j.gateway/shutdown :now 100)')
+        Clojure.eval_default('(zensols.py4j.gateway/shutdown)')