+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.0.4] - 2017-11-16
+### Changed
+- Fixed integration test.
+- Adding shutdown timeout semantics.
+- Fixed python namespace bug.
+
 ## [0.0.3] - 2017-11-16
 ### Changed
 - Better namespace handing.
@@ -25,6 +31,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Initial version
 
-[Unreleased]: https://github.com/plandes/clj-py4j/compare/v0.0.3...HEAD
+[Unreleased]: https://github.com/plandes/clj-py4j/compare/v0.0.2...HEAD
+[0.0.3]: https://github.com/plandes/clj-py4j/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/plandes/clj-py4j/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/plandes/clj-py4j/compare/v0.0.1...v0.0.2